pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
    "faker>=37.3.0",
    "pandas>=2.2.3",
]
